Modify Code refactoring of Chatbot component
